+++ src/main/java/com/polis/api/storage/Data.java
@@ -1,7 +1,5 @@
 package com.polis.api.storage;
 
-import com.polis.api.MarusiaCommand;
-
 import java.util.HashMap;
 import java.util.Map;
